+++ src/fields/ExpiredPeriodFieldset.tsx
@@ -1,21 +1,17 @@
 import React from 'react';
 import { css } from '@emotion/react';
-import { FieldBody, FieldHead, FieldSet } from './Fieldset';
+import { FieldBody, FieldHead, FieldSet } from './style';
 import CardFormLabel from 'components/card/CardFormLabel';
 import ExpiredPeriodInputContainer from 'containers/card/ExpiredPeriodInputContainer';
-
-const headStyle = css({
-  marginBottom: '3px',
-});
 
 function ExpiredPeriodFieldset() {
   return (
     <>
       <FieldSet>
-        <FieldHead style={headStyle}>
+        <FieldHead style={{ marginBottom: '10px' }}>
           <CardFormLabel>만료일</CardFormLabel>
         </FieldHead>
-        <FieldBody>
+        <FieldBody style={{ marginBottom: '20px' }}>
           <ExpiredPeriodInputContainer />
         </FieldBody>
       </FieldSet>
