+++ src/pages/CardListPage.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import CardListContainer from 'containers/card/CardListContainer';
 import PageTitle from 'components/navigater/PageTitle';
 
-import { PageWrapper, TitleWrapper } from './stlye';
+import { PageWrapper, TitleWrapper } from './style';
 
 function CardListPage() {
   return (
